Add files
The dir /model is for the purpose of handling data through database.
The dir /mode/data is for the purpose of defining what the data should be.
The dir /web-api is for the purpose of placing controllers.
The dir /utils places some tools that are used frequently.

Finishing the web-api/LoginController.java to recieve the Json from the frontend but has not be tested yet, owing to the failure building of frontend.